+++ single_cell/cmdline.py
@@ -105,6 +105,14 @@ def parse_args():
     alignment.add_argument("--bams_dir",
                            required=True,
                            help='''directory for bam storage''')
+    alignment.add_argument("--trim",
+                           default=False,
+                           action='store_true',
+                           help='''trim fastq files before aligning''')
+    alignment.add_argument("--sequencing_center",
+                           required=True,
+                           help='''sequencing center where data is generated''')
+
 
     # ===========
     # hmmcopy